company and profile validations moved to middleware
