+++ ts/controllers/Company.controller.ts
@@ -16,37 +16,11 @@ const getCompanies = async(req:Request,res:Response) => {
 
 // creates a new company
 const insertNewCompany = async(req:Request, res:Response)=>{
-    const submitData:{
-        name:string,
-        logo:string
-    } = {
-        name: req.body.name,
-        logo: req.body.logo
-    }
-    
-    if(!submitData.logo || !submitData.name){
-        return res.status(400).json({message:'Name and logo are required fields.'})
-    }
-    
-    if(typeof(submitData.name) !== 'string'
-    || typeof(submitData.logo) !== 'string'
-    ){
-        return res.status(400).json({message:'Please make sure the name and logo are of the correct type.'})
-    }
-
-    // trimming string
+    /* Validation done with middleware */
     const data:{
         name:string,
         logo:string
-    } = {
-        name: submitData.name.trim(),
-        logo: submitData.logo.trim()
-    }
-
-    // checking if there are any values left to be submitted after trimming
-    if(!data.name || !data.logo){
-        return res.status(400).json({message:'name and logo are required.'})
-    }
+    } = req.body
 
     
     try {
@@ -88,64 +62,13 @@ const getCompanyById = async (req:Request, res:Response) =>{
 
 // updating rows in the company table. Only submitted rows will be considered for update (at least one should be submitted)
 const updateCompany = async (req:Request, res:Response)=>{
-    if(!req.params.id || Number.isInteger(+req.params.id) === false){
-        return res.status(400).json({message:'Please make sure that the url parameter \'id\' is an integer.'});
-    }
+    /* Validation done with middleware */
     
-    //All three are technically optional, but something should be submitted, validation for this is 2 paragraphs lower
     const submitData:{
-        name?:string|undefined,
-        logo?:string|undefined,
-        slug?:string|undefined //added later to this object based on the submitted name
-    } = {}
-    
-    /** Validation of submitted data - checking if the submitted body parameters are not an empty string. */
-    if(req.body.name !== undefined){
-        if(req.body.name === ''){
-            return res.status(400).json({message:'There appear to be empty fields. Please check your inputs and try again.'})
-        } else {
-            submitData.name = req.body.name
-        }
-    }
-    if(req.body.logo !== undefined){
-        if(req.body.logo === ''){
-            return res.status(400).json({message:'There appear to be empty fields. Please check your inputs and try again.'})
-        } else {
-            submitData.logo = req.body.logo
-        }
-    }
-    
-    
-    // making sure at least one of the editable parameters is present
-    if(!submitData.name && !submitData.logo){
-        return res.status(400).json({message:'No data submitted for update. Please submit the fields you would like to change (logo and / or name).'})
-    }
-    
-    
-    //checking that present submitData is of correct type
-    if(
-       (submitData.name && typeof(submitData.name) !== 'string')
-    || (submitData.logo && typeof(submitData.logo) !== 'string')
-    ) {
-        return res.status(400).json({message:'Please make sure the submitted parameters are of the correct type.'})
-    }
-
-    // trimming the available inputs, and returning an error if they are an empty string after trimming
-    if(submitData.name){
-        submitData.name = submitData.name.trim()
-        if(!submitData.name){
-            return res.status(400).json({message:'Please make sure the submitted parameters are of the correct type.'})
-        }
-        submitData.slug = createSlug(submitData.name)
-    }
-
-    if(submitData.logo){
-        submitData.logo = submitData.logo.trim()
-        if(!submitData.logo){
-            return res.status(400).json({message:'Please make sure the submitted parameters are of the correct type.'})
-        }
-    }
-
+        name?:string,
+        logo?:string,
+        slug?:string
+    } = req.body
 
     // Check if company exists
     try {
